Moved x-vpc creation to vpc module
