+++ ecs_composex/common/aws.py
@@ -47,126 +47,6 @@ def get_account_id(session):
     return session.client("sts").get_caller_identity()["Account"]
 
 
-def lookup_vpc_id(session, vpc_id):
-    """
-
-    :param session: boto3 session
-    :param vpc_id: VPC ID
-    :return:
-    """
-    args = {"VpcIds": [vpc_id]}
-    arn_regexp = r"(^arn:(aws|aws-cn|aws-us-gov):ec2:([a-z]{2}-[\w]{2,6}-[0-9]{1}):([0-9]{12}):vpc\/(vpc-[a-z0-9]+)$)"
-    arn_re = re.compile(arn_regexp)
-    if vpc_id.startswith("arn:") and arn_re.match(vpc_id):
-        LOG.debug(arn_re.findall(vpc_id))
-        re_vpc_id = arn_re.findall(vpc_id)[-1][-1]
-        re_vpc_owner = arn_re.findall(vpc_id)[-1][-2]
-        args = {
-            "VpcIds": [re_vpc_id],
-            "Filters": [{"Name": "owner-id", "Values": [re_vpc_owner]}],
-        }
-        vpc_id = re_vpc_id
-    elif vpc_id.startswith("arn:") and not arn_re.match(vpc_id):
-        raise ValueError(
-            "Vpc ARN is not valid. Got", vpc_id, "Valid ARN Regexp", arn_regexp
-        )
-
-    client = session.client("ec2")
-    vpcs_r = client.describe_vpcs(**args)
-    LOG.debug(vpcs_r)
-    LOG.debug(vpcs_r["Vpcs"][0]["VpcId"])
-    if keyisset("Vpcs", vpcs_r) and vpcs_r["Vpcs"][0]["VpcId"] == vpc_id:
-        LOG.info(f"VPC Found and confirmed: {vpcs_r['Vpcs'][0]['VpcId']}")
-        return vpcs_r["Vpcs"][0]["VpcId"]
-    raise ValueError("No VPC found with ID", args["VpcIds"][0])
-
-
-def define_filter_tags(tags):
-    """
-    Function to create the filters out of tags list
-
-    :param list tags: list of Key/Value dict
-    :return: filters
-    :rtype: list
-    """
-    filters = []
-    for tag in tags:
-        key = list(tag.keys())[0]
-        filter_name = f"tag:{key}"
-        filter_values = [tag[key]]
-        filters.append({"Name": filter_name, "Values": filter_values})
-    return filters
-
-
-def lookup_vpc_from_tags(session, tags):
-    """
-    Function to find a VPC from defined Tags
-
-    :param boto3.session.Session session: boto3 session
-    :param list tags: list of tags
-    :return:
-    """
-    client = session.client("ec2")
-    filters = define_filter_tags(tags)
-    vpcs_r = client.describe_vpcs(Filters=filters)
-    if keyisset("Vpcs", vpcs_r):
-        if len(vpcs_r["Vpcs"]) > 1:
-            raise ValueError(
-                "There is more than one VPC with the provided tags.", filters
-            )
-        LOG.info(f"VPC found and confirmed: {vpcs_r['Vpcs'][0]['VpcId']}")
-        return vpcs_r["Vpcs"][0]["VpcId"]
-    raise ValueError("No VPC found with tags", filters)
-
-
-def lookup_subnets_ids(session, ids, vpc_id):
-    """
-    Function to find subnets based on a list of subnet IDs
-
-    :param session: boto3 session
-    :param ids: list of subneet IDs
-    :param str vpc_id: The VPC ID to use to search for the subnets
-    :return: list of subnets
-    :rtype: list
-    """
-    client = session.client("ec2")
-    filters = [{"Name": "vpc-id", "Values": [vpc_id]}]
-    subnets_r = client.describe_subnets(SubnetIds=ids, Filters=filters)
-    if keyisset("Subnets", subnets_r):
-        subnets = [subnet["SubnetId"] for subnet in subnets_r["Subnets"]]
-        if not all(subnet["SubnetId"] in ids for subnet in subnets_r["Subnets"]):
-            raise ValueError(
-                "Subnets returned are invalid. Expected", ids, "got", subnets
-            )
-        print(subnets, ids)
-        LOG.info(f"Subnets found and confirmed: {subnets}")
-        return subnets
-    raise ValueError("No Subnets found with provided IDs", ids)
-
-
-def lookup_subnets_from_tags(session, tags, vpc_id, subnet_key=None):
-    """
-    Function to find a VPC from defined Tags
-
-    :param boto3.session.Session session: boto3 session
-    :param list tags: list of tags
-    :param str vpc_id: The VPC ID to use to search for the subnets
-    :param str subnet_key: For troubleshooting, allows to figure which subnets this was for.
-    :return:
-    """
-    if subnet_key is None:
-        subnet_key = "Subnets "
-    client = session.client("ec2")
-    filters = define_filter_tags(tags)
-    filters.append({"Name": "vpc-id", "Values": [vpc_id]})
-    subnets_r = client.describe_subnets(Filters=filters)
-    if keyisset("Subnets", subnets_r):
-        subnets = [subnet["SubnetId"] for subnet in subnets_r["Subnets"]]
-        LOG.info(f"{subnet_key} found and confirmed: {subnets}")
-        return subnets
-    raise ValueError("No Subnets found with tags", filters)
-
-
 def assert_can_create_stack(client, name):
     """
     Checks whether a stack already exists or not
